Bugfix: Validators are now consistent.

+++ app/Http/Controllers/UserController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\User;
+use Validator;
 
 class UserController extends Controller
 {
@@ -58,13 +59,14 @@ class UserController extends Controller
             'name' => 'sometimes|required',
             'email' => 'sometimes|required|email|unique:users,email,'.$id,
             'password' => 'sometimes|required',
+            'telefone' => 'sometimes|required|regex:/^\(?\d{2}\)?\s?\d{4,5}-?\d{4}$/',
         ]);
 
         if ($validator->fails()) {
             return response()->json(['error' => $validator->errors()], 400);
         }
 
-        $data = $request->all();
+        $data = $request->only(['name', 'email', 'password', 'telefone']);
         if (isset($data['password'])) {
             $data['password'] = bcrypt($data['password']);
         }
@@ -86,6 +88,10 @@ class UserController extends Controller
         
         if (!$user) {
             return response()->json(['message' => 'User not found'], 404);
+        }
+
+        if($paciente->agendamentos()->exists()){
+            return response()->json(['error' => 'Não é possivel excluir um usuário com agendamentos'], 409);
         }
 
         $user->delete();
